screep/tests/test_requirements.txt
pytest-django==2.6.2
factory_boy
coverage
